seperate InvalidItemBehaviour tests and increase code coverage
