+++ Tests/JSONUtilitiesTests/InlineDecodingTests.swift
@@ -159,194 +159,18 @@ class InlineDecodingTests: XCTestCase {
     XCTAssert(decodedParent.children.count == 1)
   }
 
-  // MARK: Dictionary invalidItemBehaviour
+  // MARK: DecodingErrorNames
 
-  func test_stringJSONRawTypeDictionaryFails_whenThereAreInvalidObjects_and_invalidItemBehaviourIsThrow() {
-    let dictionary = [
-      "key": [
-        "key1": "value1",
-        "key2": 2
-      ]
+  func testDecodingErrorNames() {
+    let errors: [DecodingError] = [
+      .keyNotFound(dictionary: [:], key: ""),
+      .incorrectRawRepresentableRawValue(rawRepresentable: MockParent.MockEnum.self, rawValue: ""),
+      .incorrectTypeInArray(array: [], expectedType: String.self, value: 0),
+      .incorrectTypeInDictionary(dictionary: [:], key: "", expectedType: String.self, value: 0),
+      .conversionFailure(type: String.self, value: "")
     ]
-    do {
-      let _: [String: String] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .fail)
-      XCTFail("Error not thrown")
-    } catch {
-      let expectedError = DecodingError.incorrectTypeInDictionary(dictionary: dictionary, key: "key2", expectedType: String.self, value: 2)
-      XCTAssert(error as? DecodingError == expectedError)
-    }
-  }
-
-  func test_stringJSONPrimitiveConvertibleDictionaryFails_whenThereAreInvalidObjects_and_invalidItemBehaviourIsThrow() {
-    let dictionary = [
-      "key": [
-        "key1": "www.google.com",
-        "key2": 2
-      ]
-    ]
-    do {
-      let _: [String: URL] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .fail)
-      XCTFail("Error not thrown")
-    } catch {
-      let expectedError = DecodingError.incorrectTypeInDictionary(dictionary: dictionary, key: "key2", expectedType: String.self, value: 2)
-      XCTAssert(error as? DecodingError == expectedError)
-    }
-  }
-
-  func test_stringJSONObjectConvertibleDictionaryFails_whenThereAreInvalidObjects_and_invalidItemBehaviourIsThrow() {
-    let dictionary = [
-      "key": [
-        "key1": ["name": "john"],
-        "key2": 2
-      ]
-    ]
-    do {
-      let _: [String: MockSimpleChild] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .fail)
-      XCTFail("Error not thrown")
-    } catch {
-      let expectedError = DecodingError.incorrectTypeInDictionary(dictionary: dictionary, key: "key2", expectedType: JSONDictionary.self, value: 2)
-      XCTAssert(error as? DecodingError == expectedError)
-    }
-  }
-
-  func test_stringJSONRawTypeDictionary_removesInvalidObjects_invalidItemBehaviourIsRemove() {
-    let dictionary = [
-      "key": [
-        "key1": "value1",
-        "key2": 2
-      ]
-    ]
-    do {
-      let decodedDictionary: [String: String] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .remove)
-      XCTAssert(decodedDictionary.count == 1)
-    } catch {
-      XCTFail("Should not throw error")
-    }
-  }
-
-  func test_stringJSONPrimitiveConvertibleDictionary_removesInvalidObjects_invalidItemBehaviourIsRemove() {
-    let dictionary = [
-      "key": [
-        "key1": "www.google.com",
-        "key2": 2
-      ]
-    ]
-    do {
-      let decodedDictionary: [String: URL] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .remove)
-      XCTAssert(decodedDictionary.count == 1)
-    } catch {
-      XCTFail("Should not throw error")
-    }
-  }
-
-  func test_stringJSONObjectConvertibleDictionary_removesInvalidObjects_invalidItemBehaviourIsRemove() {
-    let dictionary = [
-      "key": [
-        "key1": ["name": "john"],
-        "key2": 2
-      ]
-    ]
-    do {
-      let decodedDictionary: [String: MockSimpleChild] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .remove)
-      XCTAssert(decodedDictionary.count == 1)
-    } catch {
-      XCTFail("Should not throw error")
-    }
-  }
-
-  // MARK: Array invalidItemBehaviour
-
-  func test_stringJSONRawTypeArrayFails_whenThereAreInvalidObjects_and_invalidItemBehaviourIsThrow() {
-    let dictionary = [
-      "key": [
-        "value1",
-        2
-      ]
-    ]
-    do {
-      let _: [String] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .fail)
-      XCTFail("Error not thrown")
-    } catch {
-      let expectedError = DecodingError.incorrectTypeInArray(array: [], expectedType: String.self, value: 2)
-      XCTAssert(error as? DecodingError == expectedError)
-    }
-  }
-
-  func test_stringJSONPrimitiveConvertibleArrayFails_whenThereAreInvalidObjects_and_invalidItemBehaviourIsThrow() {
-    let dictionary = [
-      "key": [
-        "www.google.com",
-        2
-      ]
-    ]
-    do {
-      let _: [URL] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .fail)
-      XCTFail("Error not thrown")
-    } catch {
-      let expectedError = DecodingError.incorrectTypeInArray(array: [], expectedType: String.self, value: 2)
-      XCTAssert(error as? DecodingError == expectedError)
-    }
-  }
-
-  func test_stringJSONObjectConvertibleArrayFails_whenThereAreInvalidObjects_and_invalidItemBehaviourIsThrow() {
-    let dictionary = [
-      "key": [
-        ["name": "john"],
-        2
-      ]
-    ]
-    do {
-      let _: [MockSimpleChild] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .fail)
-      XCTFail("Error not thrown")
-    } catch {
-      let expectedError = DecodingError.incorrectTypeInArray(array: [], expectedType: JSONDictionary.self, value: 2)
-      print(error)
-      XCTAssert(error as? DecodingError == expectedError)
-    }
-  }
-
-  func test_stringJSONRawTypeArray_removesInvalidObjects_invalidItemBehaviourIsRemove() {
-    let dictionary = [
-      "key": [
-        "value1",
-        2
-      ]
-    ]
-    do {
-      let decodedDictionary: [String] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .remove)
-      XCTAssert(decodedDictionary.count == 1)
-    } catch {
-      XCTFail("Should not throw error")
-    }
-  }
-
-  func test_stringJSONPrimitiveConvertibleArray_removesInvalidObjects_invalidItemBehaviourIsRemove() {
-    let dictionary = [
-      "key": [
-        "www.google.com",
-        2
-      ]
-    ]
-    do {
-      let decodedDictionary: [URL] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .remove)
-      XCTAssert(decodedDictionary.count == 1)
-    } catch {
-      XCTFail("Should not throw error")
-    }
-  }
-
-  func test_stringJSONObjectConvertibleArray_removesInvalidObjects_invalidItemBehaviourIsRemove() {
-    let dictionary = [
-      "key": [
-        ["name": "john"],
-        2
-      ]
-    ]
-    do {
-      let decodedDictionary: [MockSimpleChild] = try dictionary.json(atKeyPath: "key", invalidItemBehaviour: .remove)
-      XCTAssert(decodedDictionary.count == 1)
-    } catch {
-      XCTFail("Should not throw error")
+    errors.forEach {
+      XCTAssert($0.name.characters.count > 0)
     }
   }
